refactor: Clean up imports and enhance error handling in domain repository
